REFACTOR Moving the Gridsearch (#487)
